requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
alembic==1.18.4
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pydantic[email]==2.12.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
pandas==2.1.3

